Containers/FrameReduction/Container/requirements.txt
-e .
opencv-python-headless==3.4.1.15
pandas==0.23.0
pillow==5.1.0
python-dotenv>=0.5.1
requests==2.18.4